server/pyproject.toml
[tool.poetry]
name = "input-form-backend"
version = "0.1.0"
description = "Django backend for the Input Form application"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.1.0"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"

[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 